General Optimization + Mem Leak fixing + General Cleanup

+++ trabalho-pratico/includes/driver.h
@@ -19,11 +19,11 @@ char *get_driver_car_class(Driver driver);
 
 bool get_driver_account_status(Driver driver);
 
-float get_driver_average_rating (Driver driver);
+double get_driver_average_rating (Driver driver);
 
 unsigned short get_driver_total_rides (Driver driver);
 
-float get_driver_total_earned_money(Driver driver);
+double get_driver_total_earned_money(Driver driver);
 
 unsigned short get_driver_latest_ride (Driver driver);
