+++ trabalho-pratico/src/user.c
@@ -9,8 +9,7 @@
 
 #define REF_DAY "9/10/2022"
 
-typedef struct user
-{
+typedef struct user {
     char *username;
     char *name;
     char *gender;
@@ -19,19 +18,17 @@ typedef struct user
     char *pay_method;
     bool account_status;
 
-    struct user_stats
-    {
-        float average_rating;
+    struct user_stats {
+        double average_rating;
         unsigned short total_rides;
-        float total_spent_money;
+        double total_spent_money;
         unsigned short total_distance;
         unsigned short latest_ride;
         unsigned short account_age;
     } stats;
 } *User;
 
-User init_user()
-{
+User init_user() {
     User user = malloc(sizeof(struct user));
 
     user->username = NULL;
@@ -39,9 +36,9 @@ User init_user()
     user->gender = NULL;
     user->pay_method = NULL;
     user->account_status = true;
-    user->stats.average_rating = 0.0f;
+    user->stats.average_rating = 0.0;
     user->stats.total_rides = 0;
-    user->stats.total_spent_money = 0.0f;
+    user->stats.total_spent_money = 0.0;
     user->stats.total_distance = 0;
     user->stats.latest_ride = 0;
     user->stats.account_age = 0;
@@ -49,8 +46,7 @@ User init_user()
     return user;
 }
 
-User create_user(char **fields)
-{
+User create_user(char **fields) {
     User user = init_user();
 
     user->username = strdup(fields[0]);
@@ -67,70 +63,58 @@ User create_user(char **fields)
     return user;
 }
 
-char *get_user_username(User user)
-{
+char *get_user_username(User user) {
     return strdup(user->username);
 }
 
-char *get_user_name(User user)
-{
+char *get_user_name(User user) {
     return strdup(user->name);
 }
 
-char *get_user_gender(User user)
-{
+char *get_user_gender(User user) {
     return strdup(user->gender);
 }
 
-char *get_user_age(User user)
-{
+char *get_user_age(User user) {
     return get_age(user->birth_date);
 }
 
-bool get_user_account_status(User user)
-{
+bool get_user_account_status(User user) {
     return user->account_status;
 }
 
-float get_user_average_rating(User user)
-{
+double get_user_average_rating(User user) {
     return user->stats.average_rating;
 }
 
-unsigned short get_user_total_rides(User user)
-{
+unsigned short get_user_total_rides(User user) {
     return user->stats.total_rides;
 }
 
-float get_user_total_spent_money(User user)
-{
+double get_user_total_spent_money(User user) {
     return user->stats.total_spent_money;
 }
 
-unsigned short get_user_total_distance(User user)
-{
+unsigned short get_user_total_distance(User user) {
     return user->stats.total_distance;
 }
 
-unsigned short get_user_latest_ride(User user)
-{
+unsigned short get_user_latest_ride(User user) {
     return user->stats.latest_ride;
 }
 
-unsigned short get_user_account_age(User user)
-{
+unsigned short get_user_account_age(User user) {
     return user->stats.account_age;
 }
 
-void set_user_stats(User user, void **stats)
-{
+void set_user_stats(User user, void **stats) {
     unsigned short user_score = *(unsigned short *)stats[0];
-    float new_average_rating = (user->stats.average_rating * user->stats.total_rides + user_score) / (user->stats.total_rides + 1);
+    double new_average_rating = (user->stats.average_rating * user->stats.total_rides + user_score) / (user->stats.total_rides + 1);
     user->stats.average_rating = new_average_rating;
 
     user->stats.total_rides++;
 
-    float ride_cost_w_tip = *(float *)stats[1];
+    double ride_cost_w_tip = *(double *)stats[1];
     user->stats.total_spent_money += ride_cost_w_tip;
 
     unsigned short ride_distance = *(unsigned short *)stats[2];
@@ -141,8 +125,7 @@ void set_user_stats(User user, void **stats)
         user->stats.latest_ride = ride_date;
 }
 
-void free_user(User user)
-{
+void free_user(User user) {
     free(user->username);
     free(user->name);
     free(user->gender);
